# fix: removed initialization of unused object to fix innacurate benchm…